Changes for MessageAnalyzer 1.4

+++ src/MessageAnalyzerToolkit.Commands/Executables/ToolkitHelper.cs
@@ -2,44 +2,47 @@
 
 namespace MessageAnalyzerToolkit.Executables
 {
-    public static class ToolkitHelper
-    {
-        public static void DeleteETWProviderCache(string analyzerPath, string providerName)
-        {
-            string analyzerProviderName = providerName.Replace('-', '_');
+	public static class ToolkitHelper
+	{
+		public static void DeleteETWProviderCache(string analyzerPath, string providerName)
+		{
+			string analyzerProviderName = providerName.Replace('-', '_');
 
-            //delete entries
-            foreach (var opnFileName in Directory.EnumerateFiles(Path.Combine(analyzerPath, "SystemEtwManifestOPNCache"), analyzerProviderName + "*.opn", SearchOption.TopDirectoryOnly))
-            {
-                File.Delete(opnFileName);
-            }
+			//delete entries
+			foreach (var opnFileName in Directory.EnumerateFiles(Path.Combine(analyzerPath, "SystemEtwManifestOPNCache"), analyzerProviderName + "*.opn", SearchOption.TopDirectoryOnly))
+			{
+				File.Delete(opnFileName);
+			}
 
-            foreach (var opnFileName in Directory.EnumerateFiles(Path.Combine(analyzerPath, @"OPNAndConfiguration\OPNs\Facton\"), analyzerProviderName + "*.opn", SearchOption.TopDirectoryOnly))
-            {
-                File.Delete(opnFileName);
-            }
+			if (Directory.Exists(Path.Combine(analyzerPath, @"OPNAndConfiguration\OPNs\Facton\")))
+			{
+				foreach (var opnFileName in Directory.EnumerateFiles(Path.Combine(analyzerPath, @"OPNAndConfiguration\OPNs\Facton\"), analyzerProviderName + "*.opn", SearchOption.TopDirectoryOnly))
+				{
+					File.Delete(opnFileName);
+				}
+			}
 
-            foreach (var cacheFileName in Directory.EnumerateFiles(Path.Combine(analyzerPath, "CodecCache"), analyzerProviderName + "*.*", SearchOption.TopDirectoryOnly))
-            {
-                File.Delete(cacheFileName);
-            }
+			foreach (var cacheFileName in Directory.EnumerateFiles(Path.Combine(analyzerPath, "CodecCache"), analyzerProviderName + "*.*", SearchOption.TopDirectoryOnly))
+			{
+				File.Delete(cacheFileName);
+			}
 
-            foreach (var cacheFileName in Directory.EnumerateFiles(Path.Combine(analyzerPath, "CompilationCache"), analyzerProviderName + "*.*", SearchOption.TopDirectoryOnly))
-            {
-                File.Delete(cacheFileName);
-            }
-        }
+			foreach (var cacheFileName in Directory.EnumerateFiles(Path.Combine(analyzerPath, "CompilationCache"), analyzerProviderName + "*.*", SearchOption.TopDirectoryOnly))
+			{
+				File.Delete(cacheFileName);
+			}
+		}
 
-        public static void ReplaceETWProvider(string analyzerPath, string providerName, string opnFileContent)
-        {
-            string analyzerProviderName = providerName.Replace('-', '_');
+		public static void ReplaceETWProvider(string analyzerPath, string providerName, string opnFileContent)
+		{
+			string analyzerProviderName = providerName.Replace('-', '_');
 
-            DeleteETWProviderCache(analyzerPath, providerName);
+			DeleteETWProviderCache(analyzerPath, providerName);
 
-            string directoryPath = Path.Combine(analyzerPath, @"OPNAndConfiguration\OPNs\Facton\");
-            Directory.CreateDirectory(directoryPath);
+			string directoryPath = Path.Combine(analyzerPath, @"OPNAndConfiguration\OPNs\Facton\");
+			Directory.CreateDirectory(directoryPath);
 
-            File.WriteAllText(Path.Combine(directoryPath, analyzerProviderName + ".opn"), opnFileContent);
-        }
-    }
+			File.WriteAllText(Path.Combine(directoryPath, analyzerProviderName + ".opn"), opnFileContent);
+		}
+	}
 }
